style(Home): change inline classNames for tw styled component
