+++ src/pages/Home/style.ts
@@ -1,0 +1,18 @@
+import tw from 'tailwind-styled-components';
+
+export const HomeMain = tw.main`
+  max-w-screen-xl
+  mx-auto
+  mt-12
+  px-12
+  flex
+  justify-between
+`;
+
+export const ContentContainer = tw.div`
+  h-full
+  flex
+  flex-col
+  justify-around
+  items-center
+`;
